+++ test/src/Ouzo/Core/Tools/Model/Template/GeneratorTest.php
@@ -225,4 +225,9 @@ class MyImagineDialect extends Dialect
     {
         return '';
     }
+
+    protected function wrapQueryWithDistinctCount(string $sql): string
+    {
+        return $sql;
+    }
 }